feat(oracles): Created a new oracle view for the v2 oracles

+++ src/pages/Homebrew/HomebrewEditorPage/RulesSection/Stats/Stats.tsx
@@ -8,10 +8,9 @@ import {
   Typography,
 } from "@mui/material";
 import { useState } from "react";
-import { StoredRules, StoredStat } from "types/HomebrewCollection.type";
+import { StoredStat } from "types/homebrew/HomebrewRules.type";
 import { StatDialog } from "./StatDialog";
 import { useStore } from "stores/store";
-import { deleteField } from "firebase/firestore";
 import { useConfirm } from "material-ui-confirm";
 import EditIcon from "@mui/icons-material/Edit";
 import DeleteIcon from "@mui/icons-material/Delete";
@@ -19,43 +18,59 @@ import { ClampedMarkdownRenderer } from "components/shared/ClampedMarkdownRender
 
 export interface StatsProps {
   homebrewId: string;
-  stats: StoredRules["stats"];
 }
 
 export function Stats(props: StatsProps) {
-  const { homebrewId, stats } = props;
+  const { homebrewId } = props;
 
   const confirm = useConfirm();
+
+  const stats = useStore(
+    (store) => store.homebrew.collections[homebrewId]?.stats?.data ?? {}
+  );
+  const isLoading = useStore(
+    (store) => !store.homebrew.collections[homebrewId].stats?.loaded
+  );
 
   const [statDialogOpen, setStatDialogOpen] = useState(false);
   const [editingStatKey, setEditingStatKey] = useState<string | undefined>(
     undefined
   );
 
-  const updateRules = useStore((store) => store.homebrew.updateExpansionRules);
-  const addStat = (statId: string, stat: StoredStat) => {
-    return updateRules(homebrewId, { stats: { [statId]: stat } });
-  };
-  const deleteStat = (statId: string) => {
-    return updateRules(homebrewId, { stats: { [statId]: deleteField() } });
+  const addStat = useStore((store) => store.homebrew.createStat);
+  const updateStat = useStore((store) => store.homebrew.updateStat);
+  const deleteStat = useStore((store) => store.homebrew.deleteStat);
+
+  const handleStatDialogSave = (stat: StoredStat) => {
+    if (editingStatKey) {
+      return updateStat(editingStatKey, stat);
+    } else {
+      return addStat(stat);
+    }
   };
 
   const handleDelete = (statId: string) => {
-    confirm({
-      title: `Delete ${stats[statId].label}`,
-      description:
-        "Are you sure you want to delete this custom stat? It will be deleted from ALL of your custom content that references this stat.",
-      confirmationText: "Delete",
-      confirmationButtonProps: {
-        variant: "contained",
-        color: "error",
-      },
-    })
-      .then(() => {
-        deleteStat(statId).catch(() => {});
+    if (stats[statId]) {
+      confirm({
+        title: `Delete ${stats[statId].label}`,
+        description:
+          "Are you sure you want to delete this custom stat? It will be deleted from ALL of your custom content that references this stat.",
+        confirmationText: "Delete",
+        confirmationButtonProps: {
+          variant: "contained",
+          color: "error",
+        },
       })
-      .catch(() => {});
+        .then(() => {
+          deleteStat(statId).catch(() => {});
+        })
+        .catch(() => {});
+    }
   };
+
+  if (isLoading) {
+    return null;
+  }
 
   return (
     <>
@@ -66,6 +81,7 @@ export function Stats(props: StatsProps) {
           sx={{
             display: "grid",
             gridTemplateColumns: "repeat(12, 1fr)",
+            gap: 2,
             pl: 0,
             my: 0,
             listStyle: "none",
@@ -76,15 +92,18 @@ export function Stats(props: StatsProps) {
             .map((statKey) => (
               <ListItem
                 key={statKey}
-                sx={{
+                sx={(theme) => ({
                   gridColumn: { xs: "span 12", sm: "span 6", md: "span 4" },
-                }}
+                  border: `1px solid ${theme.palette.divider}`,
+                  borderRadius: 1,
+                })}
               >
                 <ListItemText
+                  secondaryTypographyProps={{ component: "span" }}
                   primary={stats[statKey].label}
                   secondary={
                     <ClampedMarkdownRenderer
-                      markdown={stats[statKey].description}
+                      markdown={stats[statKey].description ?? ""}
                       inheritColor
                     />
                   }
@@ -117,8 +136,9 @@ export function Stats(props: StatsProps) {
         Add Stat
       </Button>
       <StatDialog
+        homebrewId={homebrewId}
         open={statDialogOpen}
-        onSave={addStat}
+        onSave={handleStatDialogSave}
         onClose={() => setStatDialogOpen(false)}
         stats={stats}
         editingStatKey={editingStatKey}
